requirements.txt
django-bootstrap5
django_forms_bootstrap